requirements.txt
annotated-types==0.6.0
gunicorn==21.2.0
joblib==1.3.2
msgpack==1.0.7
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.2
redis==5.0.1
result==0.16.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
